friends list and float fixed

+++ src/main/java/ir/ac/aut/hesabgar/domain/document/UserInfo.java
@@ -76,7 +76,7 @@ public class UserInfo {
     private List<String> friendsList;
 
     @Field(OverallBalance_Col)
-    private int overallBalance;
+    private float overallBalance;
 
 //    @Field(NewInvitedGroup_Col)
 //    private List<GroupInfo> newInvitedGroup;
@@ -173,11 +173,11 @@ public class UserInfo {
         this.joinedGroupList = joinedGroupList;
     }
 
-    public int getOverallBalance() {
+    public float getOverallBalance() {
         return overallBalance;
     }
 
-    public void setOverallBalance(int overallBalance) {
+    public void setOverallBalance(float overallBalance) {
         this.overallBalance = overallBalance;
     }
